requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
transformers==4.33.2
huggingface_hub==0.16.4
scikit-learn==1.3.2
numpy
